+++ CHANGELOG.md
@@ -4,10 +4,14 @@
 
 -   **terra-form**
 <<<<<<< HEAD
+<<<<<<< HEAD
     -   form type **number**: replaced deprecated `terra-number-input` with a `mat-input` variant.
 =======
     -   **text-area** replaced deprecated `terra-text-area-input` with a `mat-input` variant.
 >>>>>>> form/terra_text_area_input
+=======
+    -   **select** replaced deprecated `terra-select-box` with a `mat-select` variant.
+>>>>>>> form/select
 
 # 11.1.0 (23.03.2021)